fix: repeated defintions in rules
