+++ _layouts/algorithm.html
@@ -4,4 +4,6 @@ layout: default
 
 {{ content }}
 
-<!--INSERT-GLOSSARY-USAGE-HERE-->
+<!-- DO NOT DELETE BELOW - This is used by a plugin to insert content -->
+<!--PLUGIN::APPEND-BACKLINK-TO-RULE-->
+<!-- DO NOT DELETE ABOVE - This is used by a plugin to insert content -->
