+++ _layouts/rule.html
@@ -83,7 +83,9 @@ layout: default
 
 {{ content }}
 
-<!--INSERT-GLOSSARY-HERE-->
+<!-- DO NOT DELETE BELOW - This is used by a plugin to insert content -->
+<!--PLUGIN::APPEND-DEFINITION-->
+<!-- DO NOT DELETE ABOVE - This is used by a plugin to insert content -->
 
 {% if page.test_aspects %}
   <h2 id="test-aspects">Test Aspects</h2>
